+++ src/renderer/src/web/main-web.ts
@@ -119,7 +119,6 @@ const validInvokeChannels: readonly string[] = [
   'measureLatency',
   'getImageDataURL',
   'getIconDataURL',
-  'getAppName',
   'changeLanguage',
   // File Share
   'getFileShareServerState',